requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
sqlalchemy==2.0.19
dbt-mysql==1.6.0
python-dotenv==1.0.0
tqdm==4.65.0